🦄 refactor(dailynote): 将 dailynote.ts 拆分为多个模块
