+++ src/func/dailynote/basic.ts
@@ -1,0 +1,65 @@
+/*
+ * Copyright (c) 2023 by Yp Z (frostime). All Rights Reserved.
+ * @Author       : Yp Z
+ * @Date         : 2023-11-12 18:06:46
+ * @FilePath     : /src/func/dailynote/basic.ts
+ * @LastEditTime : 2023-11-12 18:15:30
+ * @Description  : 
+ */
+import * as serverApi from '@/serverApi';
+import notebooks from "@/global-notebooks";
+import { info } from '@/utils';
+import { getDocsByHpath } from '@/func';
+
+/**
+ * 
+ * @param notebook 笔记本对象
+ * @returns 笔记本的 Daily Note 路径模板
+ */
+export async function getDailynoteSprig(notebookId: string): Promise<string> {
+    let conf = await serverApi.getNotebookConf(notebookId);
+    let sprig: string = conf.conf.dailyNoteSavePath;
+    return sprig;
+}
+
+/**
+ * 要求思源解析模板
+ * @param sprig
+ * @returns 
+ */
+export async function renderDailynotePath(sprig: string) {
+    return await serverApi.renderSprig(sprig);
+}
+
+
+/**
+ * 根据思源中已经有 diary 的笔记本，更新下拉框中的笔记本状态
+ * 注意，本函数不会更新 notebooks
+ * @details
+ * 1. 遍历所有笔记本，找到所有的 daily note 的 hpath
+ * 2. 对每种 hpath，调用 `await getDocsByHpath(todayDNHpath)`，查询是否存在对应的文件
+ */
+export async function currentDiaryStatus() {
+    // let todayDiary = getTodayDiaryPath();
+    //所有 hpath 的配置方案
+    let hpath_set: Set<string> = new Set();
+    notebooks.notebooks.forEach((notebook) => {
+        hpath_set.add(notebook.dailynoteHpath!);
+    });
+
+    let diaryStatus: Map<string, boolean> = new Map();
+    let count_diary = 0;
+    for (const todayDNHpath of hpath_set) {
+        //对每种 daily note 的方案，看看是否存在对应的路径
+        let docs = await getDocsByHpath(todayDNHpath);
+        if (docs.length > 0) {
+            let notebook_with_diary = docs.map(doc => doc.box);
+            notebook_with_diary.forEach((notebookId: string) => {
+                diaryStatus.set(notebookId, true);
+            });
+            count_diary += notebook_with_diary.length;
+        }
+    }
+    info(`更新日记状态: 当前日记共 ${count_diary} 篇`);
+    return diaryStatus;
+}
